requirements.txt
together
python-dotenv>=0.19.0
pandas>=1.3.0
tqdm>=4.62.0
requests>=2.26.0
